fix create symlink only if qutebrowser is installed

+++ install.sh
@@ -65,6 +65,8 @@ ln -sf "$DOTFILES/tmuxp" "$XDG_CONFIG_HOME"
 # qutebrowser #
 ###############
 
-rm -f "$XDG_CONFIG_HOME/qutebrowser/config.py"
-ln -sf "$DOTFILES/qutebrowser/config.py" "$XDG_CONFIG_HOME/qutebrowser/"
+if [ -d "$XDG_CONFIG_HOME/qutebrowser" ]; then
+  rm -f "$XDG_CONFIG_HOME/qutebrowser/config.py"
+  ln -sf "$DOTFILES/qutebrowser/config.py" "$XDG_CONFIG_HOME/qutebrowser/"
+fi
